fixed bugs

+++ dataloader.py
@@ -22,7 +22,7 @@ class DataLoader:
     def next_batch(self):
         B,T = self.B, self.T
         buf = self.tokens[self.current_position : self.current_position + B*T + 1]
-        buf = torch.tensor(buf)
+        buf = torch.tensor(buf, dtype=torch.long)
         x = (buf[:-1]).view(B,T)
         y = (buf[1:]).view(B,T)
